requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
PyJWT==2.9.0
mysql-connector-python==9.0.0
google-generativeai==0.7.2
vercel-wsgi==0.2.0